+++ src/aux-common/bots/Bot.ts
@@ -166,6 +166,10 @@ export interface BotTags {
     ['auxContextInventoryColor']?: string;
     ['auxContextInventoryHeight']?: unknown;
     ['auxContextInventoryPannable']?: boolean;
+    [`auxContextInventoryPannableMinX`]?: number | null;
+    [`auxContextInventoryPannableMaxX`]?: number | null;
+    [`auxContextInventoryPannableMinY`]?: number | null;
+    [`auxContextInventoryPannableMaxY`]?: number | null;
     ['auxContextInventoryResizable']?: boolean;
     ['auxContextInventoryRotatable']?: boolean;
     ['auxContextInventoryZoomable']?: boolean;
@@ -424,6 +428,11 @@ export const KNOWN_TAGS: string[] = [
     'auxContextInventoryHeight',
     'auxContextInventoryVisible',
     'auxContextInventoryPannable',
+    `auxContextInventoryPannableMinX`,
+    `auxContextInventoryPannableMaxX`,
+    `auxContextInventoryPannableMinY`,
+    `auxContextInventoryPannableMaxY`,
+
     'auxContextInventoryResizable',
     'auxContextInventoryRotatable',
     'auxContextInventoryZoomable',