+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changes:
 
 -   **Breaking Changes**
+
     -   Changed `create()` and `createTemp()` to automatically set `auxCreator` to the current `this` bot.
         -   `create()` no longer takes a bot/bot ID as the first parameter. Instead, you need to use the `from()` function to set the creator ID.
         -   e.g. `create(from(bot))`.
@@ -24,6 +25,7 @@
     -   Renamed `mod.subtract()` to `subtractMods()`.
     -   Renamed `mod.import()` to `getMod()`.
     -   Removed `mod.export()`.
+
 -   Improvements
     -   Added a `creator` variable to scripts and formulas which gets the bot that created the `this` bot.
         -   `creator` is null if the current bot has no creator.
@@ -37,6 +39,8 @@
         -   Should also work with bots in the `that` variable.
     -   Added a `data` variable which equals `that`.
     -   Added the `player.hideHtml()` function which hides the HTML modal.
+    -   Added in inventory tags to limit panning movements on the inventory context: `auxContextInventoryPannableMinX`, `auxContextInventoryPannableMaxX`, `auxContextInventoryPannableMinY`, `auxContextInventoryPannableMaxY`.
+    -   Reformatted new selection id logic by removing the `._` character from its return.
 
 ## V0.11.10